Move landing page and its components into page/ namespace
